@@ -2,13 +2,16 @@ package giuseppe.graziano.thermostat.model.data;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+import lombok.Data;
+
 import java.time.DayOfWeek;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.Objects;
 
 @Entity
+@Data
 public class Program {
 
     @Id
@@ -32,113 +35,4 @@ public class Program {
     @JoinColumn(name = "program_mode_id")
     private ProgramMode programMode;
 
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public boolean isActive() {
-        return active;
-    }
-
-    public void setActive(boolean active) {
-        this.active = active;
-    }
-
-    public boolean isRepete() {
-        return repete;
-    }
-
-    public void setRepete(boolean repete) {
-        this.repete = repete;
-    }
-
-    public DayOfWeek getWeekDay() {
-        return weekDay;
-    }
-
-    public void setWeekDay(DayOfWeek weekDay) {
-        this.weekDay = weekDay;
-    }
-
-    public LocalTime getStartTime() {
-        return startTime;
-    }
-
-    public void setStartTime(LocalTime startTime) {
-        this.startTime = startTime;
-    }
-
-    public LocalTime getEndTime() {
-        return endTime;
-    }
-
-    public void setEndTime(LocalTime endTime) {
-        this.endTime = endTime;
-    }
-
-    public long getSourceId() {
-        return sourceId;
-    }
-
-    public void setSourceId(long sourceId) {
-        this.sourceId = sourceId;
-    }
-
-    public boolean isSourceOn() {
-        return sourceOn;
-    }
-
-    public void setSourceOn(boolean sourceOn) {
-        this.sourceOn = sourceOn;
-    }
-
-    public ProgramMode getProgramMode() {
-        return programMode;
-    }
-
-    public void setProgramMode(ProgramMode programMode) {
-        this.programMode = programMode;
-    }
-
-    public float getTemperature() {
-        return temperature;
-    }
-
-    public void setTemperature(float temperature) {
-        this.temperature = temperature;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Program program = (Program) o;
-        return id == program.id;
-    }
-
-    @Override
-    public int hashCode() {
-
-        return Objects.hash(id);
-    }
 }
