@@ -1,7 +1,9 @@
 package giuseppe.graziano.thermostat.model.data;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 
-import javax.persistence.*;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import jakarta.persistence.*;
+import lombok.Data;
+import org.hibernate.Hibernate;
 
 import java.util.Date;
 import java.util.HashSet;
@@ -9,6 +11,7 @@ import java.util.Objects;
 import java.util.Set;
 
 @Entity
+@Data
 public class Thermostat {
 
     @JsonIgnore
@@ -65,126 +68,18 @@ public class Thermostat {
         this.mode = Thermostat.MANUAL_MODE;
     }
 
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public boolean isActive() {
-        return active;
-    }
-
-    public void setActive(boolean active) {
-        this.active = active;
-    }
-
-    public boolean isStateOn() {
-        return stateOn;
-    }
-
-    public void setStateOn(boolean stateOn) {
-        this.stateOn = stateOn;
-    }
-
-    public Date getLastMeasurement() {
-        return lastMeasurement;
-    }
-
-    public void setLastMeasurement(Date lastMeasurement) {
-        this.lastMeasurement = lastMeasurement;
-    }
-
-    public float getTemperature() {
-        return temperature;
-    }
-
-    public void setTemperature(float temperature) {
-        this.temperature = temperature;
-    }
-
-    public Set<Sensor> getSensors() {
-        return sensors;
-    }
-
-    public void setSensors(Set<Sensor> sensors) {
-        this.sensors = sensors;
-    }
-
-    public ProgramMode getProgramMode() {
-        return programMode;
-    }
-
-    public void setProgramMode(ProgramMode programMode) {
-        this.programMode = programMode;
-    }
-
-    public Set<Source> getSources() {
-        return sources;
-    }
-
-    public void setSources(Set<Source> sources) {
-        this.sources = sources;
-    }
-
-    public String getMode() {
-        return mode;
-    }
-
-    public void setMode(String mode) {
-        this.mode = mode;
-    }
-
-    public ManualMode getManualMode() {
-        return manualMode;
-    }
-
-    public void setManualMode(ManualMode manualMode) {
-        this.manualMode = manualMode;
-    }
-
-    public Source getSource() {
-        return source;
-    }
-
-    public void setSource(Source source) {
-        this.source = source;
-    }
-
-    public boolean isCold() {
-        return cold;
-    }
-
-    public void setCold(boolean cold) {
-        this.cold = cold;
-    }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
         Thermostat that = (Thermostat) o;
-        return id == that.id;
+        return Objects.equals(getId(), that.getId());
     }
 
+    @Override
+    public int hashCode() {
+        return getClass().hashCode();
+    }
 }
 
 
