@@ -3,20 +3,12 @@ package giuseppe.graziano.thermostat.model.data;
 
 import java.util.Set;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.JoinTable;
-import javax.persistence.ManyToMany;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
+import jakarta.persistence.*;
+import lombok.Data;
 
 @Entity
 @Table(name = "user_table")
+@Data
 public class User {
 
     @Id
@@ -36,57 +28,4 @@ public class User {
     @JoinTable(name = "users_thermostats", joinColumns = @JoinColumn(name = "user_id", referencedColumnName = "id"), inverseJoinColumns = @JoinColumn(name = "thermostat_id", referencedColumnName = "id"))
     private Set<Thermostat> thermostats;
 
-
-    public User() {
-        super();
-    }
-
-    //
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public boolean isAdmin() {
-        return isAdmin;
-    }
-
-    public Long getSelectedThermostatId() {
-        return selectedThermostatId;
-    }
-
-    public void setSelectedThermostatId(Long selectedThermostatId) {
-        this.selectedThermostatId = selectedThermostatId;
-    }
-
-    public void setAdmin(boolean admin) {
-        isAdmin = admin;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public Set<Thermostat> getThermostats() {
-        return thermostats;
-    }
-
-    public void setThermostats(Set<Thermostat> thermostats) {
-        this.thermostats = thermostats;
-    }
 }
